fix(route-docs): keep schema loading client-side
